refactor: Convert configuration from TOML to JSON format - Replace TOML with JSON for configuration files - Create dedicated config.py module for configuration handling - Move load_config() and ensure_config_exists() to config.py - Update tools.py to import from config module - Create config_template.json with same structure as TOML version - Remove config_template.toml file - Remove toml dependency, use native json module instead - This improves maintainability and removes external dependencies
